refactor(config/presets): split parsePreset to parse.ts
